tp: improve build rules for amalgamating SQL files
Until now, we've been relying on GN variables + a home-brewn GN action
for generating the amalgamted SQL files for metrics. However, this is
barely holding together today and with the advent of
the standard library this is only going to get worse.

Fix the problem by introducing new `perfetto_sql_source_set` and
`perfetto_sql_amalgamated_header` GN templates which model the
"source_set" and "amalgamation" operations respectively. To prove it
works, we migrate the initial version of the standard library to this
new APIs.

There is some spillover into gen* scripts but ended up being reasonably
straightforward. Metrics will be migrated in a followup CL.

Bug: 255535171
Change-Id: I864a1a3df2fc8d9bb0b26dc2a943476e9c68722a